11/3

+++ frmPatientRecord.cs
@@ -83,45 +83,52 @@ namespace Capstone
             displayPrescription(ODSPH, ODCYL, ODAXIS, ODADD, ODPD, OSSPH, OSCYL, OSAXIS, OSADD, OSPD);
             if (colName == "EditPRecord")
             {
-                frmAddPatientRecord frm = new frmAddPatientRecord(this);
-                frm.txtAddFName.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[1].Value.ToString();
-                frm.txtAddLName.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[2].Value.ToString();
-                frm.txtAddAddress.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[5].Value.ToString();
-                frm.txtAddContact.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[6].Value.ToString();
-                frm.comBoxAddGender.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[8].Value.ToString();
-                frm.txtAddAge.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[7].Value.ToString();
-                frm.txtAddPatientID.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[3].Value.ToString();
-                
-
-                frm.txtAddODSPH.Text = txtODSPH.Text;
-                frm.txtADDODCYL.Text = txtODCYL.Text;
-                frm.txtAddODAXIS.Text = txtODAXIS.Text;
-                frm.txtAddODADD.Text = txtODADD.Text;
-                frm.txtAddODPD.Text = txtODPD.Text;
-
-                frm.txtAddOSSPH.Text = txtOSSPH.Text;
-                frm.txtADDOSCYL.Text = txtOSCYL.Text;
-                frm.txtAddOSAXIS.Text = txtOSAXIS.Text;
-                frm.txtAddOSADD.Text = txtOSADD.Text;
-                frm.txtAddOSPD.Text = txtOSPD.Text;
-
-
-                frm.ShowDialog();
-
-            }else if (colName == "del")
-            {
-                if (MessageBox.Show("Remove this Item?", title, MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                frmPermission frmP = new frmPermission();
+                frmP.ShowDialog();
+                if (frmP.lblGrant.Text == "1")
                 {
-                    cn.Open();
-                    cm = new SqlCommand("DELETE FROM tblPatientRecord WHERE Patient_ID LIKE '" + PatID + "'", cn);
-                    cm.ExecuteNonQuery();
-                    cn.Close();
-                    MessageBox.Show("Record has been successfully deleted.", title, MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    frmAddPatientRecord frm = new frmAddPatientRecord(this);
+                    frm.txtAddFName.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[1].Value.ToString();
+                    frm.txtAddLName.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[2].Value.ToString();
+                    frm.txtAddAddress.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[5].Value.ToString();
+                    frm.txtAddContact.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[6].Value.ToString();
+                    frm.comBoxAddGender.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[8].Value.ToString();
+                    frm.txtAddAge.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[7].Value.ToString();
+                    frm.txtAddPatientID.Text = dataGridViewPatientRecord.Rows[e.RowIndex].Cells[3].Value.ToString();
 
-                    //LoadStock();
-                    classLoadData.LoadRecordsPatient(dataGridViewPatientRecord, txtSearchPatientRecord);
+                    frm.txtAddODSPH.Text = txtODSPH.Text;
+                    frm.txtADDODCYL.Text = txtODCYL.Text;
+                    frm.txtAddODAXIS.Text = txtODAXIS.Text;
+                    frm.txtAddODADD.Text = txtODADD.Text;
+                    frm.txtAddODPD.Text = txtODPD.Text;
+
+                    frm.txtAddOSSPH.Text = txtOSSPH.Text;
+                    frm.txtADDOSCYL.Text = txtOSCYL.Text;
+                    frm.txtAddOSAXIS.Text = txtOSAXIS.Text;
+                    frm.txtAddOSADD.Text = txtOSADD.Text;
+                    frm.txtAddOSPD.Text = txtOSPD.Text;
+
+                    frm.ShowDialog();
                 }
-                
+            }
+            else if (colName == "del")
+            {
+                frmPermission frmP = new frmPermission();
+                frmP.ShowDialog();
+                if (frmP.lblGrant.Text == "1")
+                {
+                    if (MessageBox.Show("Remove this Item?", title, MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                    {
+                        cn.Open();
+                        cm = new SqlCommand("DELETE FROM tblPatientRecord WHERE Patient_ID LIKE '" + PatID + "'", cn);
+                        cm.ExecuteNonQuery();
+                        cn.Close();
+                        MessageBox.Show("Record has been successfully deleted.", title, MessageBoxButtons.OK, MessageBoxIcon.Information);
+
+                        //LoadStock();
+                        classLoadData.LoadRecordsPatient(dataGridViewPatientRecord, txtSearchPatientRecord);
+                    }
+                }
             }
             else
             {
